[nimble] Making sun compiler happy
git-svn-id: https://gridkit.googlecode.com/svn/grid-lab/trunk/nimble@2288 4e96dbfe-9890-11de-ac5d-73b9d092dc0a

+++ probe-jmx/src/main/java/org/gridkit/lab/util/jmx/mxstruct/coherence/ClusterMXStruct.java
@@ -24,12 +24,12 @@ public class ClusterMXStruct extends MXStruct {
 
 	@AttrName("LocalMemberId")
 	public int getLocalMemberId() {
-		return super.getMXAttr();
+		return (Integer)super.getMXAttr();
 	}
 
 	@AttrName("ClusterSize")
 	public int getClusterSize() {
-		return super.getMXAttr();
+		return (Integer)super.getMXAttr();
 	}
 
 	@AttrName("LicenseMode")
@@ -54,12 +54,12 @@ public class ClusterMXStruct extends MXStruct {
 
 	@AttrName("MembersDepartureCount")
 	public long getMembersDepartureCount() {
-		return super.getMXAttr();
+		return (Long)super.getMXAttr();
 	}
 
 	@AttrName("OldestMemberId")
 	public int getOldestMemberId() {
-		return super.getMXAttr();
+		return (Integer)super.getMXAttr();
 	}
 
 	@AttrName("Version")
